# Documentation/swagger.yaml
openapi: 3.0.0
info:
  title: Authentication API
  description: API for user authentication including Google OAuth and user registration.
  version: 1.0.0
servers:
  - url: http://localhost:3000
    description: Local server
...
paths:
...
  /api/shorten:
    post:
      tags:
        - URL Shortening
      summary: Create a shortened URL
      operationId: createShortUrl
      description: >
        Create a shortened URL for the provided long URL. The user must be authenticated through Google OAuth and must not exceed the rate limit for URL creation.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateShortUrlDTO'
      responses:
        '201':
          description: URL shortened successfully
          content:
            application/json:
              example:
                shortUrl: 'http://short.ly/abc123'
                createdAt: '2024-11-24T10:00:00Z'
        '400':
          description: Invalid request data
        '429':
          description: Rate limit exceeded
  /api/shorten/{alias}:
    get:
      tags:
        - URL Shortening
      summary: Redirect to long URL
      operationId: redirectToLongUrl
      parameters:
        - name: alias
          in: path
          required: true
          description: The alias for the shortened URL
          schema:
            type: string
      responses:
        '302':
          description: Redirects to the long URL
        '401':
          description: Unauthorized access - Google OAuth required
          content:
            application/json:
              example:
                statusCode: 401
                message: "Unauthorized"
        '404':
          description: Short URL not found
...
components:
  schemas:
...
    CreateShortUrlDTO:
      type: object
      properties:
        longUrl:
          type: string
          format: url
          description: The original long URL to be shortened
          example: 'https://www.example.com/some/long/url'
        customAlias:
          type: string
          description: A custom alias for the shortened URL (optional)
          example: 'myCustomAlias'
        topic:
          type: string
          description: A topic or category associated with the URL (optional)
          example: 'technology'
      required:
        - longUrl
  securitySchemes:
    GoogleAuth:
      type: oauth2
      flows:
        authorizationCode:
          authorizationUrl: 'https://accounts.google.com/o/oauth2/auth'
          tokenUrl: 'https://oauth2.googleapis.com/token'
          scopes:
            'profile': 'Access user profile information'
            'email': 'Access user email'
security:
  - GoogleAuth: []